requirements.txt
streamlit
google-generativeai
streamlit-mic-recorder
reportlab
pandas
Pillow